Extract old AST to TechTalk.SpecFlow.Parser.Compatibility
