+++ Parser.Compatibility/packages.config
@@ -1,0 +1,4 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="Gherkin" version="3.1.2" targetFramework="net45" />
+</packages>
