+++ TechTalk.SpecFlow.sln
@@ -1,7 +1,6 @@
-
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 2013
-VisualStudioVersion = 12.0.31101.0
+# Visual Studio 14
+VisualStudioVersion = 14.0.23107.0
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Setup", "Setup", "{DCE0C3C4-5BC6-4A30-86BE-3FEFF4677A01}"
 EndProject
@@ -56,6 +55,8 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = ".nuget", ".nuget", "{CB927B
 	ProjectSection(SolutionItems) = preProject
 		.nuget\packages.config = .nuget\packages.config
 	EndProjectSection
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TechTalk.SpecFlow.Parser.Compatibility", "Parser.Compatibility\TechTalk.SpecFlow.Parser.Compatibility.csproj", "{B4588E81-C589-434F-8F12-3D881433C3E4}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -147,6 +148,12 @@ Global
 		{33B3107E-666D-4C31-96C0-745697E51973}.Release|Any CPU.Build.0 = Release|Any CPU
 		{33B3107E-666D-4C31-96C0-745697E51973}.VS2010IntegrationTest|Any CPU.ActiveCfg = Debug|Any CPU
 		{33B3107E-666D-4C31-96C0-745697E51973}.VS2010IntegrationTest|Any CPU.Build.0 = Debug|Any CPU
+		{B4588E81-C589-434F-8F12-3D881433C3E4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B4588E81-C589-434F-8F12-3D881433C3E4}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B4588E81-C589-434F-8F12-3D881433C3E4}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B4588E81-C589-434F-8F12-3D881433C3E4}.Release|Any CPU.Build.0 = Release|Any CPU
+		{B4588E81-C589-434F-8F12-3D881433C3E4}.VS2010IntegrationTest|Any CPU.ActiveCfg = Release|Any CPU
+		{B4588E81-C589-434F-8F12-3D881433C3E4}.VS2010IntegrationTest|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
