+++ Tests/TechTalk.SpecFlow.Specs/TechTalk.SpecFlow.Specs.csproj
@@ -295,6 +295,10 @@
       <Project>{27B6695E-7049-4C1B-8820-812722128D9E}</Project>
       <Name>SpecFlowBinPackage</Name>
     </ProjectReference>
+    <ProjectReference Include="..\..\Parser.Compatibility\TechTalk.SpecFlow.Parser.Compatibility.csproj">
+      <Project>{b4588e81-c589-434f-8f12-3d881433c3e4}</Project>
+      <Name>TechTalk.SpecFlow.Parser.Compatibility</Name>
+    </ProjectReference>
     <ProjectReference Include="..\..\Parser\TechTalk.SpecFlow.Parser.csproj">
       <Project>{7CCEF6D6-FC17-422E-9BED-EDD752B6496F}</Project>
       <Name>TechTalk.SpecFlow.Parser</Name>
